refactor dispenser module
